txmgr: Remove goroutine use in queue test
Assertions should be made from the main test routine, if possible.
I had local test failures where assertions where made after the main
routine returned already. No idea if this was related to changes in this
PR.

A waitgroup can be used here alternatively, but it's cleaner to just
assert in the main routine.